requirements.txt
docker
gunicorn==23.0.0
joblib
mkl-fft
mkl-random
mkl-service
mlflow
numpy
optuna
pandas
pip
python-dotenv
scikit-learn
setuptools
wheel
xgboost
fastapi
uvicorn
pydantic
psutil
